CAP 2 EX 3
=> Implement INNER mode of the algorithm

+++ Miguel/Exercicios/Capitulo_2/ex3_media/middleware.c
@@ -9,9 +9,9 @@ void clear_buffer() {
 }
 
 
-void show_average(int average) {
+void show_average(float average) {
     
-    printf("O valor médio dos números digitados é %d\n\n", average);
+    printf("O valor médio dos números digitados é %.2f\n\n", average);
     return;
 }
