+++ Miguel/Exercicios/Capitulo_2/ex3_media/middleware.h
@@ -3,6 +3,6 @@
 
 void clear_buffer();
 
-void show_average(int average);
+void show_average(float average);
 
 int acummulator();
